Move more out of utils, clear unused
