+++ app/lib/features/bug_report/pages/bug_report_page.dart
@@ -13,6 +13,16 @@ import 'package:logging/logging.dart';
 
 final _log = Logger('a3::bug_report');
 
+String? _getIssueId(String url) {
+  // example - https://github.com/bitfriend/acter-bugs/issues/9
+  RegExp re = RegExp(r'^https:\/\/github.com\/(.*)\/(.*)\/issues\/(\d*)$');
+  RegExpMatch? match = re.firstMatch(url);
+  if (match != null) {
+    return match.group(3);
+  }
+  return null;
+}
+
 class BugReportPage extends ConsumerStatefulWidget {
   static const titleField = Key('bug-report-title');
   static const includeScreenshot = Key('bug-report-include-screenshot');
@@ -73,7 +83,7 @@ class _BugReportState extends ConsumerState<BugReportPage> {
         screenshotPath: withScreenshot ? widget.imagePath : null,
         extraFields: extraFields,
       );
-      String? issueId = getIssueId(reportUrl);
+      String? issueId = _getIssueId(reportUrl);
       loadingNotifier.update((state) => false);
       if (context.mounted) {
         final status = issueId != null
